fix bug where adding a second user would push/pop the add user view + reorganization
