+++ ntfy/Views/Settings/SettingsView.swift
@@ -1,0 +1,59 @@
+import Foundation
+import SwiftUI
+import StoreKit
+
+struct SettingsView: View {
+    @EnvironmentObject private var store: Store
+    @State private var userDialog: UserDialog?
+    
+    var body: some View {
+        NavigationView {
+            Form {
+                Section(
+                    header: Text("General"),
+                    footer: Text("When subscribing to new topics, this server will be used as a default.")
+                ) {
+                    DefaultServerView()
+                }
+                Section(
+                    header: Text("Users"),
+                    footer: Text("To access read-protected topics, you may add or edit users here. All topics for a given server will use the same user.")
+                ) {
+                    UserTableView(dialog: $userDialog)
+                }
+                Section(header: Text("About")) {
+                    AboutView()
+                }
+            }
+            .navigationTitle("Settings")
+        }
+        .sheet(item: $userDialog) { dialog in
+            UserEditorView(
+                selectedUser: dialog.user,
+                onSave: { baseUrl, username, password in
+                    store.saveUser(baseUrl: baseUrl, username: username, password: password)
+                    userDialog = nil
+                },
+                onDelete: { user in
+                    store.delete(user: user)
+                    userDialog = nil
+                },
+                onCancel: {
+                    userDialog = nil
+                }
+            )
+        }
+        .navigationViewStyle(StackNavigationViewStyle())
+    }
+}
+
+
+struct SettingsView_Previews: PreviewProvider {
+    static var previews: some View {
+        let store = Store.preview // Store.previewEmpty
+        SettingsView()
+            .environment(\.managedObjectContext, store.context)
+            .environmentObject(store)
+            .environmentObject(AppDelegate())
+    }
+}
